+++ plugins/org.eclipse.n4js/src/org/eclipse/n4js/validation/validators/N4JSFlowgraphValidator.java
@@ -17,8 +17,8 @@ import org.eclipse.n4js.postprocessing.ASTFlowInfo;
 import org.eclipse.n4js.projectModel.IN4JSCore;
 import org.eclipse.n4js.resource.N4JSResource;
 import org.eclipse.n4js.smith.Measurement;
+import org.eclipse.n4js.smith.N4JSDataCollectors;
 import org.eclipse.n4js.utils.FindReferenceHelper;
-import org.eclipse.n4js.utils.N4JSDataCollectors;
 import org.eclipse.n4js.validation.AbstractN4JSDeclarativeValidator;
 import org.eclipse.n4js.validation.validators.flowgraphs.DeadCodeValidator;
 import org.eclipse.n4js.validation.validators.flowgraphs.FlowValidator;
